requirements.txt
asgiref==3.4.1
Django==3.2.5
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
